\name{kcm}
\docType{data}
\alias{kcm}
\title{
Kentucky yearly cancer mortality from 1999-2005.
}
\description{This table gives the yearly mortality counts due to neoplasms (ICD 10 codes C00-D48) and population sizes for Kentucky from 1999-2005.  For more information, see http://wonder.cdc.gov/wonder/help/cmf.html.}
\usage{data(kcm)}
\format{A 7 by 3 data frame.}
\source{
Centers for Disease Control and Prevention, National Center for Health Statistics. Compressed Mortality File 1999-2005. CDC WONDER On-line Database, compiled from Compressed Mortality File 1999-2005 Series 20 No. 2K, 2008.  Accessed at
http://wonder.cdc.gov/cmf-icd10.html on May 5, 2008.
}
\keyword{datasets}
